remove web support (#23)
Too much work to keep both web and native devices up to sync. Not only just graphical differences but also in terms of the threading model they use as well as for the `big_double` package which can produce bogus results due to the invariance of `int` and `double` on web versus native. 

Just a headache for now, has a chance of making a comeback later on if the future model allows for it. For now, out of the question.

Signed-off-by: Jiaming Meng <jackmeng0814@gmail.com>